KJHH-1246: Refactoring & add organisaatiohakutype
