+++ src/main/static/src/actions/actiontypes.js
@@ -175,6 +175,10 @@ export const FETCH_ALL_ORGANISAATIOS_REQUEST = "FETCH_ALL_ORGANISAATIOS_REQUEST"
 export const FETCH_ALL_ORGANISAATIOS_SUCCESS = "FETCH_ALL_ORGANISAATIOS_SUCCESS";
 export const FETCH_ALL_ORGANISAATIOS_FAILURE = "FETCH_ALL_ORGANISAATIOS_FAILURE";
 
+export const FETCH_ALL_ORGANISAATIOS_HIERARCHY_REQUEST = "FETCH_ALL_ORGANISAATIOS_HIERARCHY_REQUEST";
+export const FETCH_ALL_ORGANISAATIOS_HIERARCHY_SUCCESS = "FETCH_ALL_ORGANISAATIOS_HIERARCHY_SUCCESS";
+export const FETCH_ALL_ORGANISAATIOS_HIERARCHY_FAILURE = "FETCH_ALL_ORGANISAATIOS_HIERARCHY_FAILURE";
+
 export const FETCH_ALL_RYHMAT_REQUEST = "FETCH_ALL_RYHMAT_REQUEST";
 export const FETCH_ALL_RYHMAT_SUCCESS = "FETCH_ALL_RYHMAT_SUCCESS";
 export const FETCH_ALL_RYHMAT_FAILURE = "FETCH_ALL_RYHMAT_FAILURE";
