+++ src/main/static/src/components/common/henkilo/labelvalues/LinkitetytHenkilot.js
@@ -10,7 +10,7 @@ import type {HenkiloState} from "../../../../reducers/henkilo.reducer";
 import type {L} from "../../../../types/localisation.type";
 import {unlinkHenkilo, fetchHenkiloSlaves} from '../../../../actions/henkilo.actions';
 import type { KayttooikeusOrganisaatiot } from '../../../../types/domain/kayttooikeus/KayttooikeusPerustiedot.types'
-import { hasAnyPalveluRooli } from '../../../../utilities/organisaatio.util'
+import { hasAnyPalveluRooli } from '../../../../utilities/palvelurooli.util'
 
 type Props = {
     kayttooikeudet: Array<KayttooikeusOrganisaatiot>,
