+++ src/main/static/src/types/domain/organisaatio/organisaatio.type.js
@@ -1,0 +1,27 @@
+// @flow
+
+export type OrganisaatioHakuTulos = {
+    numHits: number,
+    organisaatiot: Array<Organisaatio>
+}
+
+export type Organisaatio = {
+    oid: string,
+    alkuPvm: string,
+    lakkautusPvm: string,
+    parentOid: string,
+    parentOidPath: string,
+    ytunnus: string,
+    oppilaitosKoodi: string,
+    oppilaitostyyppi: string,
+    toimipistekoodi: string,
+    match: boolean,
+    nimi: {[string]: string },
+    kieletUris: Array<string>,
+    kotipaikkaUri: string,
+    children: Array<Organisaatio>,
+    aliOrganisaatioMaara: number,
+    virastoTunnus: string,
+    organisaatiotyypit: Array<string>,
+    status: string
+}
